# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investment Timing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Asset category
 */
export type StockDataCategory =
  (typeof StockDataCategory)[keyof typeof StockDataCategory];

export const StockDataCategory = {
  "us-stocks": "us-stocks",
  "kr-stocks": "kr-stocks",
  indices: "indices",
} as const;

/**
 * RSI signal - buy if < 30, sell if >= 70, neutral otherwise
 */
export type StockDataRsiSignal =
  (typeof StockDataRsiSignal)[keyof typeof StockDataRsiSignal];

export const StockDataRsiSignal = {
  buy: "buy",
  sell: "sell",
  neutral: "neutral",
} as const;

export interface PricePoint {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Closing price */
  close: number;
  /** MA20 at this date */
  ma20: number;
  /** Deviation from MA20 in percent */
  deviationPercent: number;
  /** Upper Bollinger Band */
  bbUpper: number;
  /** Middle Bollinger Band (SMA20) */
  bbMiddle: number;
  /** Lower Bollinger Band */
  bbLower: number;
  /** MACD line value */
  macdLine: number;
  /** MACD signal line value */
  signalLine: number;
  /** MACD histogram value */
  macdHistogram: number;
}

export interface StockData {
  /** Stock ticker symbol (Yahoo Finance format) */
  symbol: string;
  /** Short display symbol for UI */
  displaySymbol: string;
  /** Full company/fund name */
  name: string;
  /** Asset category */
  category: StockDataCategory;
  /** Current stock price */
  currentPrice: number;
  /** Previous day closing price */
  previousClose: number;
  /** Price change percentage from previous close */
  changePercent: number;
  /** RSI(14) value (0-100) */
  rsi14: number;
  /** RSI signal - buy if < 30, sell if >= 70, neutral otherwise */
  rsiSignal: StockDataRsiSignal;
  /** 20-day moving average */
  ma20: number;
  /** Current price deviation from MA20 in percent */
  ma20DeviationPercent: number;
  /** Bollinger Band upper line (SMA20 + 2*stddev) */
  bbUpper: number;
  /** Bollinger Band middle line (SMA20) */
  bbMiddle: number;
  /** Bollinger Band lower line (SMA20 - 2*stddev) */
  bbLower: number;
  /** True when current price is at or below the lower Bollinger Band */
  isTouchingLowerBand: boolean;
  /** True when price touches lower BB AND RSI < 30 — extreme buy opportunity */
  isSuperBuySignal: boolean;
  /** MACD line value (EMA12 - EMA26) */
  macdLine: number;
  /** MACD signal line (EMA9 of MACD) */
  signalLine: number;
  /** MACD histogram (MACD - Signal) */
  macdHistogram: number;
  /** Historical closing prices (last 30 days with all indicators) */
  historicalPrices: PricePoint[];
  /** ISO timestamp of last update */
  lastUpdated: string;
  /** Trading volume */
  volume: number;
  /** Currency code (USD, KRW, etc.) */
  currency: string;
}

export interface StocksResponse {
  stocks: StockData[];
  /** ISO timestamp of last update */
  lastUpdated: string;
  /** List of symbols currently triggering the super buy alert */
  superBuySignals: string[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}