# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Stock Investment Timing Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stocks
    description: Stock data and indicators
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stocks:
    get:
      operationId: getStocks
      tags: [stocks]
      summary: Get stock data with indicators
      description: Returns current price, RSI(14), MA20, MACD, Bollinger Bands for all tracked symbols
      responses:
        "200":
          description: Stock data with technical indicators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StocksResponse"
  /stocks/{symbol}:
    get:
      operationId: getStock
      tags: [stocks]
      summary: Get single stock data
      description: Returns stock data for a specific symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock data with technical indicators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockData"
        "404":
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StockData:
      type: object
      properties:
        symbol:
          type: string
          description: Stock ticker symbol (Yahoo Finance format)
        displaySymbol:
          type: string
          description: Short display symbol for UI
        name:
          type: string
          description: Full company/fund name
        category:
          type: string
          enum: [us-stocks, kr-stocks, indices]
          description: Asset category
        currentPrice:
          type: number
          description: Current stock price
        previousClose:
          type: number
          description: Previous day closing price
        changePercent:
          type: number
          description: Price change percentage from previous close
        rsi14:
          type: number
          description: RSI(14) value (0-100)
        rsiSignal:
          type: string
          enum: [buy, sell, neutral]
          description: RSI signal - buy if < 30, sell if >= 70, neutral otherwise
        ma20:
          type: number
          description: 20-day moving average
        ma20DeviationPercent:
          type: number
          description: Current price deviation from MA20 in percent
        bbUpper:
          type: number
          description: Bollinger Band upper line (SMA20 + 2*stddev)
        bbMiddle:
          type: number
          description: Bollinger Band middle line (SMA20)
        bbLower:
          type: number
          description: Bollinger Band lower line (SMA20 - 2*stddev)
        isTouchingLowerBand:
          type: boolean
          description: True when current price is at or below the lower Bollinger Band
        isSuperBuySignal:
          type: boolean
          description: True when price touches lower BB AND RSI < 30 — extreme buy opportunity
        macdLine:
          type: number
          description: MACD line value (EMA12 - EMA26)
        signalLine:
          type: number
          description: MACD signal line (EMA9 of MACD)
        macdHistogram:
          type: number
          description: MACD histogram (MACD - Signal)
        historicalPrices:
          type: array
          description: Historical closing prices (last 30 days with all indicators)
          items:
            $ref: "#/components/schemas/PricePoint"
        lastUpdated:
          type: string
          description: ISO timestamp of last update
        volume:
          type: number
          description: Trading volume
        currency:
          type: string
          description: Currency code (USD, KRW, etc.)
      required:
        - symbol
        - displaySymbol
        - name
        - category
        - currentPrice
        - previousClose
        - changePercent
        - rsi14
        - rsiSignal
        - ma20
        - ma20DeviationPercent
        - bbUpper
        - bbMiddle
        - bbLower
        - isTouchingLowerBand
        - isSuperBuySignal
        - macdLine
        - signalLine
        - macdHistogram
        - historicalPrices
        - lastUpdated
        - volume
        - currency
    PricePoint:
      type: object
      properties:
        date:
          type: string
          description: Date in YYYY-MM-DD format
        close:
          type: number
          description: Closing price
        ma20:
          type: number
          description: MA20 at this date
        deviationPercent:
          type: number
          description: Deviation from MA20 in percent
        bbUpper:
          type: number
          description: Upper Bollinger Band
        bbMiddle:
          type: number
          description: Middle Bollinger Band (SMA20)
        bbLower:
          type: number
          description: Lower Bollinger Band
        macdLine:
          type: number
          description: MACD line value
        signalLine:
          type: number
          description: MACD signal line value
        macdHistogram:
          type: number
          description: MACD histogram value
      required:
        - date
        - close
        - ma20
        - deviationPercent
        - bbUpper
        - bbMiddle
        - bbLower
        - macdLine
        - signalLine
        - macdHistogram
    StocksResponse:
      type: object
      properties:
        stocks:
          type: array
          items:
            $ref: "#/components/schemas/StockData"
        lastUpdated:
          type: string
          description: ISO timestamp of last update
        superBuySignals:
          type: array
          description: List of symbols currently triggering the super buy alert
          items:
            type: string
      required:
        - stocks
        - lastUpdated
        - superBuySignals
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message