# lib/api-zod/src/generated/types/pricePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investment Timing Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface PricePoint {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Closing price */
  close: number;
  /** MA20 at this date */
  ma20: number;
  /** Deviation from MA20 in percent */
  deviationPercent: number;
  /** Upper Bollinger Band */
  bbUpper: number;
  /** Middle Bollinger Band (SMA20) */
  bbMiddle: number;
  /** Lower Bollinger Band */
  bbLower: number;
  /** MACD line value */
  macdLine: number;
  /** MACD signal line value */
  signalLine: number;
  /** MACD histogram value */
  macdHistogram: number;
}