# lib/api-zod/src/generated/types/stockData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investment Timing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PricePoint } from "./pricePoint";
import type { StockDataCategory } from "./stockDataCategory";
import type { StockDataRsiSignal } from "./stockDataRsiSignal";

export interface StockData {
  /** Stock ticker symbol (Yahoo Finance format) */
  symbol: string;
  /** Short display symbol for UI */
  displaySymbol: string;
  /** Full company/fund name */
  name: string;
  /** Asset category */
  category: StockDataCategory;
  /** Current stock price */
  currentPrice: number;
  /** Previous day closing price */
  previousClose: number;
  /** Price change percentage from previous close */
  changePercent: number;
  /** RSI(14) value (0-100) */
  rsi14: number;
  /** RSI signal - buy if < 30, sell if >= 70, neutral otherwise */
  rsiSignal: StockDataRsiSignal;
  /** 20-day moving average */
  ma20: number;
  /** Current price deviation from MA20 in percent */
  ma20DeviationPercent: number;
  /** Bollinger Band upper line (SMA20 + 2*stddev) */
  bbUpper: number;
  /** Bollinger Band middle line (SMA20) */
  bbMiddle: number;
  /** Bollinger Band lower line (SMA20 - 2*stddev) */
  bbLower: number;
  /** True when current price is at or below the lower Bollinger Band */
  isTouchingLowerBand: boolean;
  /** True when price touches lower BB AND RSI < 30 — extreme buy opportunity */
  isSuperBuySignal: boolean;
  /** MACD line value (EMA12 - EMA26) */
  macdLine: number;
  /** MACD signal line (EMA9 of MACD) */
  signalLine: number;
  /** MACD histogram (MACD - Signal) */
  macdHistogram: number;
  /** Historical closing prices (last 30 days with all indicators) */
  historicalPrices: PricePoint[];
  /** ISO timestamp of last update */
  lastUpdated: string;
  /** Trading volume */
  volume: number;
  /** Currency code (USD, KRW, etc.) */
  currency: string;
}