# lib/api-zod/src/generated/types/stocksResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stock Investment Timing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { StockData } from "./stockData";

export interface StocksResponse {
  stocks: StockData[];
  /** ISO timestamp of last update */
  lastUpdated: string;
  /** List of symbols currently triggering the super buy alert */
  superBuySignals: string[];
}